refactor(upload-dialog): componentize
